Streamline server.js with separated routes
