+++ db/connection.js
@@ -1,0 +1,10 @@
+const mysql = require("mysql2");
+
+const db = mysql.createConnection({
+  host: "localhost",
+  user: "root",
+  password: "N0fish!!",
+  database: "election",
+});
+
+module.exports = db;
