+++ routes/apiRoutes/index.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+
+router.use(require("./candidateRoutes"));
+router.use(require("./partyRoutes"));
+
+module.exports = router;
