+++ server.js
@@ -1,195 +1,29 @@
 const express = require("express");
 const { json } = require("express/lib/response");
-const mysql = require("mysql2");
 const inputCheck = require("./utils/inputCheck");
+const db = require("./db/connection");
+
+const apiRoutes = require("./routes/apiRoutes");
 
 // Middleware
 const PORT = process.env.PORT || 3001;
 const app = express();
 
-// Database connection
-const db = mysql.createConnection(
-  {
-    host: "localhost",
-    user: "root",
-    password: "N0fish!!",
-    database: "election",
-  },
-  console.log("Connected to the election database.")
-);
-
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
 
-// GET all candidates
-app.get("/api/candidates", (req, res) => {
-  const sql = `SELECT candidates.*, parties.name 
-  AS party_name 
-  FROM candidates 
-  LEFT JOIN parties 
-  ON candidates.party_id = parties.id`;
+app.use("/api", apiRoutes);
 
-  db.query(sql, (err, rows) => {
-    if (err) {
-      res.status(500).json({ error: err.message });
-      return;
-    }
-    res.json({ message: "success", data: rows });
-  });
-});
-
-// GET a single candidate
-app.get("/api/candidate/:id", (req, res) => {
-  const sql = `SELECT candidates.*, parties.name 
-  AS party_name 
-  FROM candidates 
-  LEFT JOIN parties 
-  ON candidates.party_id = parties.id 
-  WHERE candidates.id = ?`;
-  const params = [req.params.id];
-  db.query(sql, params, (err, row) => {
-    if (err) {
-      res.status(400).json({ error: err.message });
-      return;
-    }
-    res.json({
-      message: "success",
-      data: row,
-    });
-  });
-});
-
-// DELETE a candidate
-app.delete("/api/candidate/:id", (req, res) => {
-  const sql = `DELETE FROM candidates WHERE id = ?`;
-  const params = [req.params.id];
-
-  db.query(sql, params, (err, result) => {
-    if (err) {
-      res.status(400).json({ error: err.message });
-      return;
-    } else if (!result.affectedRows) {
-      res.json({
-        message: "Candidate not found",
-      });
-    } else {
-      res.json({
-        message: "deleted",
-        changes: result.affectedRows,
-        id: req.params.id,
-      });
-    }
-  });
-});
-
-// CREATE a candidate
-app.post("/api/candidate", ({ body }, res) => {
-  const errors = inputCheck(
-    body,
-    "first_name",
-    "last_name",
-    "industry_connected"
-  );
-  if (errors) {
-    res.status(400).json({ error: errors });
-    return;
-  }
-
-  const sql = `INSERT INTO candidates (first_name, last_name, industry_connected)
-    VALUES (?,?,?)`;
-  const params = [body.first_name, body.last_name, body.industry_connected];
-
-  db.query(sql, params, (err, result) => {
-    if (err) {
-      res.status(400).json({ error: err.message });
-      return;
-    }
-    res.json({
-      message: "success",
-      data: body,
-    });
-  });
-});
-
-// PUT update a candidates party
-app.put("/api/candidate/:id", (req, res) => {
-  const errors = inputCheck(req.body, "party_id");
-
-  if (errors) {
-    res.status(400).json({ error: errors });
-    return;
-  }
-  const sql = `UPDATE candidates SET party_id = ? WHERE id = ?`;
-  const params = [req.body.party_id, req.params.id];
-  db.query(sql, params, (err, result) => {
-    if (err) {
-      res.status(400).json({ error: err.message });
-    } else if (!result.affectedRows) {
-      res.json({ message: "Candidate not found" });
-    } else {
-      res.json({
-        message: "success",
-        data: req.body,
-        change: result.affectedRows,
-      });
-    }
-  });
-});
-
-// GET all parties
-app.get("/api/parties", (req, res) => {
-  const sql = `SELECT * FROM parties`;
-  db.query(sql, (err, rows) => {
-    if (err) {
-      res.status(500).json({ error: err.message });
-      return;
-    }
-    res.json({
-      message: "success",
-      data: rows,
-    });
-  });
-});
-
-// GET party by ID
-app.get("/api/party/:id", (req, res) => {
-  const sql = `SELECT * FROM parties WHERE id = ?`;
-  params = [req.params.id];
-  db.query(sql, params, (err, row) => {
-    if (err) {
-      res.status(400).json({ error: err.message });
-      return;
-    }
-    res.json({
-      message: "success",
-      data: row,
-    });
-  });
-});
-
-// DELETE party by id
-app.delete("/api/party/:id", (req, res) => {
-  const sql = `DELETE FROM parties WHERE id = ?`;
-  params = [req.params.id];
-  db.query(sql, params, (err, result) => {
-    if (err) {
-      res.status(400).json({ error: err.message });
-    } else if (!result.affectedRows) {
-      res.json({ message: "Party not found" });
-    } else {
-      res.json({
-        message: "deleted",
-        changes: result.affectedRows,
-        id: req.params.id,
-      });
-    }
-  });
-});
-
+// Default response for any other request (Not Found)
 app.use((req, res) => {
   res.status(404).end();
 });
 
-app.listen(PORT, () => {
-  console.log(`Server running on http://localhost:${PORT}`);
+// Start server after DB connection
+db.connect((err) => {
+  if (err) throw err;
+  console.log("Database connected.");
+  app.listen(PORT, () => {
+    console.log(`Server running on http://localhost:${PORT}`);
+  });
 });
